refactor: split ServerMultiplayerRoomMixin into composable focused mixins
